Converted the comparison buffer to uint

+++ Source/UnrealGPUSwarm/GPUBitonicSort.cpp
@@ -22,7 +22,7 @@ public:
 		SHADER_PARAMETER(FIntVector, job_params)
 		SHADER_PARAMETER(uint32_t, itemCount) // the number of particles
 
-		SHADER_PARAMETER_UAV(StructuredBuffer<float>, comparisonBuffer)
+		SHADER_PARAMETER_UAV(StructuredBuffer<uint>, comparisonBuffer)
 		SHADER_PARAMETER_UAV(RWStructuredBuffer<uint>, indexBuffer)
 	END_SHADER_PARAMETER_STRUCT()
 
@@ -49,7 +49,7 @@ public:
 		SHADER_PARAMETER(FIntVector, job_params)
 		SHADER_PARAMETER(uint32_t, itemCount) // the number of particles
 
-		SHADER_PARAMETER_UAV(StructuredBuffer<float>, comparisonBuffer)
+		SHADER_PARAMETER_UAV(StructuredBuffer<uint>, comparisonBuffer)
 		SHADER_PARAMETER_UAV(RWStructuredBuffer<uint>, indexBuffer)
 	END_SHADER_PARAMETER_STRUCT()
 
@@ -77,7 +77,7 @@ public:
 		SHADER_PARAMETER(FIntVector, job_params)
 		SHADER_PARAMETER(uint32_t, itemCount) // the number of particles
 
-		SHADER_PARAMETER_UAV(StructuredBuffer<float>, comparisonBuffer)
+		SHADER_PARAMETER_UAV(StructuredBuffer<uint>, comparisonBuffer)
 		SHADER_PARAMETER_UAV(RWStructuredBuffer<uint>, indexBuffer)
 	END_SHADER_PARAMETER_STRUCT()
